+++ darts-hub/control/Updater.cs
@@ -21,7 +21,7 @@ namespace darts_hub.control
         // ATTRIBUTES
 
         // Increase for new build ..
-        public static readonly string version = "b1.2.11";
+        public static readonly string version = "b1.2.12";
         
 
         public static event EventHandler<ReleaseEventArgs>? NoNewReleaseFound;